Extract constants, helpers, indicators, themes
